+++ src/main/java/com/daniel/kakeapp/Income/IncomeController.java
@@ -6,6 +6,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -33,7 +34,7 @@ public class IncomeController {
 
 @GetMapping("/totalIncomeByMonth")
     @ResponseBody
-    public Double totalIncomeByMonth(@RequestParam int month) {
+    public BigDecimal totalIncomeByMonth(@RequestParam int month) {
         return incomeService.findTotalIncomesByMonth(month);
     }
 }
